Add files via upload
The goal of this project is to revamp and redesign the email templates , with a focus on improving the visual appeal, engagement, and user experience. These templates are an essential part of our communication with users, and we want to ensure that they not only convey information effectively but also provide a pleasant and engaging interaction.